style(windows): silence final MSVC C4305/C4244 warnings (3rd pass) (#24)
Cleans up the ~31 sites left after #15. Each fix is mechanical — adds
an explicit cast to match the destination type — except for two cases
where the chained-assignment expression was restructured into separate
statements so the assignment's `_Ty` result is no longer being narrowed
to `int`.

Sites by file:
  - src/f2dock/Docking.cpp                 12
  - src/PG-range/PG.cc                      6
  - src/fft-utils/fastfft.cpp               3
  - src/fast-hydro/pseudoGsol.cpp           3
  - src/fast-PQ/PairingHeap.cpp             1  (chain split)
  - src/GB-rerank/pairingHeap.cpp           1  (chain split)
  - src/PG-range/cuckoo.cc                  1  (srand cast)
  - src/fft-utils/rank-fftw.cpp             1  (srand cast)
  - src/f2dock/F2Dock.cpp                   1  (srand cast)
  - src/f3dock-loop-closure/sturm.h         1  (ptrdiff_t -> int)
  - src/f2dock/TopValues.cpp                1  (m_CurMin VAL_TYPE -> int)

After this pass the Windows Debug build should be free of all
non-libcvc C4305/C4244 conversion warnings.

libcvc warnings (`build/_deps/libcvc-src/...`) remain untouched and
will be addressed upstream.

+++ src/PG-range/PG.cc
@@ -279,9 +279,9 @@ void PG::addPoint(Point *a) {
   center.z += (float)(TRANSLATE);
   //  double radius = a->getr();
 
-  int cz = (int)center.z / DIM;
-  int cy = (int)center.y / DIM;
-  int cx = (int)center.x / DIM;
+  int cz = (int)((int)center.z / DIM);
+  int cy = (int)((int)center.y / DIM);
+  int cx = (int)((int)center.x / DIM);
   RangeTuple<plane *> temp;
   plane *p;
   line *l;
@@ -339,9 +339,9 @@ void PG::removePoint(Point *a) {
   //  double radius = a->getr();
 
   cout << "REMOVING AN ATOM " << endl;
-  int cz = (int)center.z / DIM;
-  int cy = (int)center.y / DIM;
-  int cx = (int)center.x / DIM;
+  int cz = (int)((int)center.z / DIM);
+  int cy = (int)((int)center.y / DIM);
+  int cx = (int)((int)center.x / DIM);
 
   vector<RangeTuple<plane *>> P = g.RR.report(cz, cz);
   if (P.empty()) {
